+++ CHANGELOG.md
@@ -6,7 +6,8 @@
 
 - Adds `QueryX.Count` (#175)
 - World stats contain separate reserved/used memory fields (#177)
-- Provides `MapX` for up to 12 components (#182) 
+- Provides `MapX` for up to 12 components (#182)
+- Adds `MapX.Set` and `Map.Set` (#183)
 
 ### Documentation